+++ navigation/lib-navigation-observer/src/main/java/ru/surfstudio/android/navigation/observer/navigator/activity/ActivityNavigatorWithResultFactory.kt
@@ -2,15 +2,20 @@ package ru.surfstudio.android.navigation.observer.navigator.activity
 
 import androidx.appcompat.app.AppCompatActivity
 import ru.surfstudio.android.navigation.navigator.activity.ActivityNavigatorFactory
+import ru.surfstudio.android.navigation.observer.ScreenResultEmitter
+import ru.surfstudio.android.navigation.observer.storage.RouteStorage
 
 /**
  * Factory to create [ActivityNavigatorWithResultImpl]
  */
-open class ActivityNavigatorWithResultFactory : ActivityNavigatorFactory() {
+open class ActivityNavigatorWithResultFactory(
+    private val screenResultEmitter: ScreenResultEmitter,
+    private val routeStorage: RouteStorage
+) : ActivityNavigatorFactory() {
 
     override fun create(
-            parentActivity: AppCompatActivity
+        parentActivity: AppCompatActivity
     ): ActivityNavigatorWithResult {
-        return ActivityNavigatorWithResultImpl(parentActivity)
+        return ActivityNavigatorWithResultImpl(parentActivity, screenResultEmitter, routeStorage)
     }
 }