+++ navigation/lib-navigation-observer/src/main/java/ru/surfstudio/android/navigation/observer/route/PermissionRequestRoute.kt
@@ -1,0 +1,16 @@
+package ru.surfstudio.android.navigation.observer.route
+
+import ru.surfstudio.android.navigation.route.activity.ActivityRoute
+
+/**
+ * Route for requesting permission
+ * This is used with [RequestPermission] command.
+ */
+class PermissionRequestRoute(
+    val permissions: Array<String>
+) : ActivityRoute(), ResultRoute<Boolean> {
+
+    override fun getId(): String {
+        return "PermissionRequestRoute"
+    }
+}