+++ navigation/sample-standard/src/main/java/ru/surfstudio/android/navigation/sample_standard/screen/bottom_nav/BottomNavFragmentView.kt
@@ -7,7 +7,6 @@ import android.view.ViewGroup
 import com.jakewharton.rxbinding2.view.clicks
 import kotlinx.android.synthetic.main.fragment_bottom_nav.*
 import ru.surfstudio.android.core.mvp.binding.rx.ui.BaseRxFragmentView
-import ru.surfstudio.android.navigation.provider.container.FragmentNavigationContainer
 import ru.surfstudio.android.navigation.provider.container.TabFragmentNavigationContainer
 import ru.surfstudio.android.navigation.sample_standard.R
 import ru.surfstudio.android.navigation.sample_standard.utils.addOnBackPressedListener
@@ -26,10 +25,9 @@ class BottomNavFragmentView : BaseRxFragmentView(), TabFragmentNavigationContain
         return inflater.inflate(R.layout.fragment_bottom_nav, container, false)
     }
 
-    override fun onActivityCreated(savedInstanceState: Bundle?, viewRecreated: Boolean) {
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?, viewRecreated: Boolean) {
         addOnBackPressedListener { bm.backPressed.accept() }
         home_tab_btn.clicks().bindTo { bm.bottomNavClicked.accept(BottomNavTabType.HOME) }
         profile_tab_btn.clicks().bindTo { bm.bottomNavClicked.accept(BottomNavTabType.PROFILE) }
     }
-
 }