+++ push/sample/src/main/java/ru/surfstudio/android/push/sample/ui/common/notification/strategies/simple/DataPushStrategy.kt
@@ -45,7 +45,7 @@ class DataPushStrategy : BaseSimplePushStrategy<NotificationTypeData>() {
             context.applicationContext,
             title.hashCode(),
             intent,
-            PendingIntent.FLAG_ONE_SHOT
+            getPendingIntentFlags()
         )
     }