+++ shared-pref/sample/src/main/java/ru/surfstudio/android/shared/pref/sample/app/dagger/CustomAppComponent.kt
@@ -5,6 +5,7 @@ import ru.surfstudio.android.dagger.scope.PerApplication
 import ru.surfstudio.android.sample.dagger.app.dagger.DefaultAppComponent
 import ru.surfstudio.android.sample.dagger.app.dagger.DefaultAppModule
 import ru.surfstudio.android.sample.dagger.app.dagger.DefaultSharedPrefModule
+import ru.surfstudio.android.sample.dagger.app.dagger.NavigationModule
 import ru.surfstudio.android.shared.pref.sample.interactor.common.network.EtagModule
 import ru.surfstudio.android.shared.pref.sample.interactor.common.network.NetworkModule
 import ru.surfstudio.android.shared.pref.sample.interactor.common.network.OkHttpModule
@@ -15,15 +16,18 @@ import ru.surfstudio.android.shared.pref.sample.interactor.ip.IpRepository
 import ru.surfstudio.android.shared.pref.sample.interactor.ip.cache.IpStorage
 
 @PerApplication
-@Component(modules = [
-    DefaultAppModule::class,
-    DefaultSharedPrefModule::class,
-    CacheModule::class,
-    EtagModule::class,
-    NetworkModule::class,
-    OkHttpModule::class,
-    ServerUrlModule::class,
-    IpModule::class])
+@Component(
+    modules = [
+        DefaultAppModule::class,
+        DefaultSharedPrefModule::class,
+        NavigationModule::class,
+        CacheModule::class,
+        EtagModule::class,
+        NetworkModule::class,
+        OkHttpModule::class,
+        ServerUrlModule::class,
+        IpModule::class]
+)
 interface CustomAppComponent : DefaultAppComponent {
     fun ipRepository(): IpRepository
     fun ipStorage(): IpStorage