+++ android-standard-version-plugin/RELEASE_NOTES.md
@@ -4,6 +4,7 @@
 
 ## 0.5.0-alpha
 ##### Android-standard-version-plugin
+* ANDDEP-1259 Update versions
 * ANDDEP-521 Added gradle plugin for providing artifacts versions
 * ANDDEP-982 `kotlin` version raised: `1.3.31 -> 1.3.71`;
 * ANDDEP-982 `gson` version raised: `2.8.5 -> 2.8.6`;