+++ mvi/RELEASE_NOTES.md
@@ -1,8 +1,12 @@
 # Core-mvi Release Notes
 
-- [0.5.0-alpha](#050-alpha)
+- [0.6.0-alpha](#060-alpha)
+- [0.5.0](#050)
 
-## 0.5.0-alpha
+## 0.6.0-alpha
+##### Core-mvi
+* TODO
+## 0.5.0
 ##### Core-mvi
 * ANDDEP-1008 License added
 * ANDDEP-671 Core mvi refactor, add comments