+++ shared-pref/RELEASE_NOTES.md
@@ -1,8 +1,12 @@
 # Shared-pref Release Notes
 
-- [0.5.0-alpha](#050-alpha)
+- [0.6.0-alpha](#060-alpha)
+- [0.5.0](#050)
 
-## 0.5.0-alpha
+## 0.6.0-alpha
+##### Shared-pref
+* TODO
+## 0.5.0
 ##### Shared-pref
 * ANDDEP-687 Changed "com.annimon:stream" dependency from "api" to "implementation" type
 * ANDDEP-782 one Editor instance per SP, added optional prefs name, remove, clear + get/put for float and double methods, default values for all, optional 'async' flag